Refactor http connector
