--- Sources/ProcessOut/Sources/Connectors/HttpAuth/HttpConnectorAuthCredentials.swift
@@ -1,15 +1,0 @@
-//
-//  HttpConnectorAuthCredentials.swift
-//  ProcessOut
-//
-//  Created by Andrii Vysotskyi on 31.03.2023.
-//
-
-struct HttpConnectorAuthCredentials {
-
-    /// Project id to associate requests with.
-    let projectId: String
-
-    /// Project's private key.
-    let privateKey: String?
-}
