+++ Sources/ProcessOut/Sources/Core/Extensions/URLSessionTask+Cancellable.swift
@@ -1,8 +1,8 @@
 //
-//  URLSessionTask+Extensions.swift
+//  URLSessionTask+Cancellable.swift
 //  ProcessOut
 //
-//  Created by Andrii Vysotskyi on 21.12.2022.
+//  Created by Andrii Vysotskyi on 03.04.2023.
 //
 
 import Foundation
